Updated the Ada example to use Alire.
Alire supports various linuxes, windows and macOS (I'm not sure about
arch linux, work is definitely in progress).

I put myself in as maintainer in alire.toml; they do require an email
address for the 'maintainers' table, but since it's quite unlikely (?)
we'll want to release an Alire crate for this example it may not
matter.

You may want to move main.adb and the executable back to ada/.

  * ada/.gitignore: ignore alire-related directories.
  * ada/README.md: updated for Alire.
  * ada/alire.toml: new.
  * ada/main.gpr: added Alire-related options.
  * ada/src/main.adb: renamed from ada/main.adb. Tidied up compiler
      warnings generated in Alire's 'development' profile.