Add CI for ada/.
Uses alire to do the builds, in job 'alire'.

Includes a temporary job, 'alire-macos-13', which builds ada/ using
macOS 13 and the latest stable Xcode. This is to check for possible
effects of the Xcode 15 change to 'ld', which has been found to
interfere with Ada exception handling. There isn't any exception
handling in this code, so there _should_ be no impact.

  * .github/workflows/main.yml
    (alire): new. Builds ada/ for windows-latest, ubuntu-latest,
      macos-latest.
    (alire-macos-13): new. Builds ada/ for macos-13/xcode-latest.